refactor(slash_commands): merge llm_profile into slash_commands
llm_profile was only consumed by slash_commands (/model and /provider
commands). Merged into a single file to consolidate LLM profile command
handling. No logic changes.
